Extract `schedule_early_close` tests
